Vitis-4626 Deprecated all XCL APIs
Move context of include/xrt.h to include/deprecated/xrt.h

The APIs are deprecated and are for internal use only.
The replacement APIs are xrt native C++ APIs.

Post deprecation the appropriate APIs will move to a private none
exported header.
